+++ companion/lib/Controls/ControlTypes/PageNumber.ts
@@ -10,23 +10,10 @@ import {
 	ButtonGraphicsDecorationType,
 	ButtonGraphicsElementUsage,
 	ButtonGraphicsShowStatusIcons,
-	type DrawStyleLayeredButtonModel,
 } from '@companion-app/shared/Model/StyleModel.js'
-import { ConvertSomeButtonGraphicsElementForDrawing } from '../../Graphics/ConvertGraphicsElements.js'
-import { ElementConversionCache } from '../../Graphics/ElementConversionCache.js'
-import { ControlBase } from '../ControlBase.js'
 import type { ControlDependencies } from '../ControlDependencies.js'
-import type {
-	ControlWithoutActions,
-	ControlWithoutActionSets,
-	ControlWithoutEvents,
-	ControlWithoutLayeredStyle,
-	ControlWithoutOptions,
-	ControlWithoutPushed,
-} from '../IControlFragments.js'
 import { CreateElementOfType } from './Button/LayerDefaults.js'
-
-const emptyMap: ReadonlyMap<string, never> = new Map<string, never>()
+import { ControlButtonPage } from './PageButton.js'
 
 export const pageNumberElements: SomeButtonGraphicsElement[] = [
 	{
@@ -97,35 +84,8 @@ export const pageNumberElements: SomeButtonGraphicsElement[] = [
  * Individual Contributor License Agreement for Companion along with
  * this program.
  */
-export class ControlButtonPageNumber
-	extends ControlBase<PageNumberButtonModel>
-	implements
-		ControlWithoutActions,
-		ControlWithoutLayeredStyle,
-		ControlWithoutEvents,
-		ControlWithoutActionSets,
-		ControlWithoutOptions,
-		ControlWithoutPushed
-{
+export class ControlButtonPageNumber extends ControlButtonPage<PageNumberButtonModel> {
 	readonly type = 'pagenum'
-
-	readonly supportsActions = false
-	readonly supportsEntities = false
-	readonly supportsLayeredStyle = false
-	readonly supportsEvents = false
-	readonly supportsActionSets = false
-	readonly supportsOptions = false
-	readonly supportsPushed = false
-
-	/**
-	 * The variables referenced in the last draw. Whenever one of these changes, a redraw should be performed
-	 */
-	#lastDrawVariables: ReadonlySet<string> | null = null
-
-	/**
-	 * Cache for element conversion results (for future per-element caching optimization)
-	 */
-	readonly #elementConversionCache = new ElementConversionCache()
 
 	/**
 	 * @param registry - the application core
@@ -151,81 +111,11 @@ export class ControlButtonPageNumber
 		}
 	}
 
-	/**
-	 * Prepare this control for deletion
-	 */
-	destroy(): void {
-		this.#elementConversionCache.clear()
-		super.destroy()
-	}
-
-	/**
-	 * Get the complete style object of a button
-	 * @returns the processed style of the button
-	 */
-	#lastDrawStyle: DrawStyleLayeredButtonModel | null = null
-	getLastDrawStyle(): DrawStyleLayeredButtonModel | null {
-		return this.#lastDrawStyle
-	}
-
-	/**
-	 * Get the complete style object of a button
-	 * @returns the processed style of the button
-	 */
-	async getDrawStyle(): Promise<DrawStyleLayeredButtonModel | null> {
-		const location = this.deps.pageStore.getLocationOfControlId(this.controlId)
-		const parser = this.deps.variableValues.createVariablesAndExpressionParser(location, null, null)
-
-		// Compute the new drawing, using the element conversion cache for per-element caching
-		const { elements, usedVariables } = await ConvertSomeButtonGraphicsElementForDrawing(
-			this.deps.instance.definitions,
-			parser,
-			this.deps.graphics.renderPixelBuffers.bind(this.deps.graphics),
-			pageNumberElements,
-			emptyMap,
-			true,
-			this.#elementConversionCache
-		)
-		this.#lastDrawVariables = usedVariables.size > 0 ? usedVariables : null
-
-		const result: DrawStyleLayeredButtonModel = {
-			pushed: false,
-			stepCurrent: 0,
-			stepCount: 0,
-			button_status: undefined,
-			action_running: undefined,
-
-			elements,
-
-			style: 'button-layered',
+	protected getDrawElements(): ReturnType<ControlButtonPage<any>['getDrawElements']> {
+		return {
 			drawType: 'pagenum',
+			elements: pageNumberElements,
 		}
-
-		this.#lastDrawStyle = result
-		return result
-	}
-
-	/**
-	 * Collect the connection ids, labels, and variables referenced by this control
-	 * @param foundConnectionIds - connection ids being referenced
-	 * @param foundConnectionLabels - connection labels being referenced
-	 * @param foundVariables - variables being referenced
-	 */
-	collectReferencedConnectionsAndVariables(
-		_foundConnectionIds: Set<string>,
-		_foundConnectionLabels: Set<string>,
-		_foundVariables: Set<string>
-	): void {
-		// Nothing being referenced
-	}
-
-	/**
-	 * Inform the control that it has been moved, and anything relying on its location must be invalidated
-	 */
-	triggerLocationHasChanged(): void {
-		// Ensure any dependencies on the location in the drawing are updated
-		this.#elementConversionCache.clear()
-		this.triggerRedraw()
 	}
 
 	/**
@@ -248,23 +138,5 @@ export class ControlButtonPageNumber
 		return {
 			type: this.type,
 		}
-	}
-
-	renameVariables(_labelFrom: string, _labelTo: string): void {
-		// Nothing to do
-	}
-	/**
-	 * Propagate variable changes
-	 * @param allChangedVariables - variables with changes
-	 */
-	onVariablesChanged(allChangedVariables: ReadonlySet<string>): void {
-		if (!this.#lastDrawVariables) return
-		if (this.#lastDrawVariables.isDisjointFrom(allChangedVariables)) return
-
-		// Queue invalidation for cached elements that use any of the changed variables
-		this.#elementConversionCache.queueInvalidateVariables(allChangedVariables)
-
-		this.logger.silly('variable changed in button ' + this.controlId)
-		this.triggerRedraw()
 	}
 }
